+++ .github/workflows/upm-subtree-split.yml
@@ -4,19 +4,13 @@ on:
   push:
     branches:
     - main
+
 jobs:
   upm-subtree-split:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
       with:
         fetch-depth: 0
-    - name: upm subtree split
-      run: |
-        $packageDir = Get-Item -Path "**/Packages/com.*" | Select-Object -ExpandProperty FullName
-        $packageDir = $packageDir.replace('${{ github.workspace }}/','')
-        Write-Host $packageDir
-        git subtree split --prefix="$packageDir" -b upm
-        git push -u origin upm --force
-      working-directory: ${{ github.workspace }}
-      shell: pwsh
+
+    - uses: RageAgainstThePixel/upm-subtree-split@v1
