--- Utilities.Async/Assets/Demo.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c1efb0f37b2705c458c02abdac776e35
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
