+++ Utilities.Async/Packages/com.utilities.async/Runtime/Async/ActionExtensions.cs
@@ -1,0 +1,16 @@
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System;
+
+namespace Utilities.Async
+{
+    public static class ActionExtensions
+    {
+        /// <summary>
+        /// Invokes the <see cref="Action"/> on the unity main thread.
+        /// </summary>
+        /// <param name="action">The action to invoke.</param>
+        public static void InvokeOnMainThread(this Action action)
+            => AwaiterExtensions.RunOnUnityScheduler(action);
+    }
+}
