+++ Utilities.Async/Packages/com.utilities.async/Samples~/Utilities.Async.Samples.asmdef
@@ -1,6 +1,6 @@
 {
-    "name": "Utilities.Async.Demo",
-    "rootNamespace": "Utilities.Async.Demo",
+    "name": "Utilities.Async.Samples",
+    "rootNamespace": "Utilities.Async.Samples",
     "references": [
         "GUID:a6609af893242c7438d701ddd4cce46a"
     ],
