+++ Utilities.Async/Packages/manifest.json
@@ -1,9 +1,9 @@
 {
   "dependencies": {
-    "com.unity.addressables": "1.21.18",
-    "com.unity.ide.rider": "3.0.25",
-    "com.unity.ide.visualstudio": "2.0.21",
-    "com.utilities.buildpipeline": "1.1.8"
+    "com.unity.addressables": "1.21.19",
+    "com.unity.ide.rider": "3.0.27",
+    "com.unity.ide.visualstudio": "2.0.22",
+    "com.utilities.buildpipeline": "1.1.9"
   },
   "scopedRegistries": [
     {
