+++ Utilities.Async/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2022.3.10f1
-m_EditorVersionWithRevision: 2022.3.10f1 (ff3792e53c62)
+m_EditorVersion: 2022.3.16f1
+m_EditorVersionWithRevision: 2022.3.16f1 (d2c21f0ef2f1)
